refactor(activerecord): remove PG adapter re-export shim, fix all imports
Delete adapters/postgresql-adapter.ts shim. Update all imports
(tests, tasks, connection-adapters, test-adapter, trailties, package.json)
to reference the canonical path at connection-adapters/postgresql-adapter.ts.
