+++ packages/activerecord/src/adapters/index.ts
@@ -1,3 +1,3 @@
 export { SQLite3Adapter } from "../connection-adapters/sqlite3-adapter.js";
-export { PostgreSQLAdapter } from "./postgresql-adapter.js";
+export { PostgreSQLAdapter } from "../connection-adapters/postgresql-adapter.js";
 export { Mysql2Adapter } from "./mysql2-adapter.js";
