--- packages/activerecord/src/adapters/postgresql-adapter.ts
@@ -1,7 +1,0 @@
-// Re-export from canonical location for backwards compatibility.
-export {
-  PostgreSQLAdapter,
-  StatementPool,
-  MoneyDecoder,
-} from "../connection-adapters/postgresql-adapter.js";
-export type { IndexDefinition } from "../connection-adapters/postgresql-adapter.js";
