+++ packages/activerecord/src/establish-connection.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeEach, afterEach } from "vitest";
 import { Base } from "./base.js";
-import { PostgreSQLAdapter } from "./adapters/postgresql-adapter.js";
+import { PostgreSQLAdapter } from "./connection-adapters/postgresql-adapter.js";
 import { SQLite3Adapter } from "./connection-adapters/sqlite3-adapter.js";
 import { Mysql2Adapter } from "./adapters/mysql2-adapter.js";
 import { ConnectionHandler } from "./connection-adapters/abstract/connection-handler.js";
